# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger for Gocure API",
        "title": "Gocure API",
        "contact": {
            "name": "Gocure Support",
            "email": "rodrigo@odhin.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7087",
    "basePath": "/",
    "paths": {
        "/embed/toFeature": {
            "post": {
                "description": "embed files to feature of a cucumber report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "embed"
                ],
                "summary": "embed files to feature",
                "parameters": [
                    {
                        "description": "Embed Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Embed"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    }
                }
            }
        },
        "/embed/toScenario": {
            "post": {
                "description": "embed files to scenario of a cucumber report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "embed"
                ],
                "summary": "embed files to scenario",
                "parameters": [
                    {
                        "description": "Embed Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Embed"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    }
                }
            }
        },
        "/embed/toStep": {
            "post": {
                "description": "embed files to step of a cucumber report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "embed"
                ],
                "summary": "embed files to step",
                "parameters": [
                    {
                        "description": "Embed Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Embed"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    }
                }
            }
        },
        "/html/generate": {
            "post": {
                "description": "generate html reports from cucumber report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "html"
                ],
                "summary": "generate html reports",
                "parameters": [
                    {
                        "description": "HTML Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HTML"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "handlers.ResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/fiber.Map"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "models.Embed": {
            "type": "object",
            "properties": {
                "featureIndex": {
                    "type": "integer"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inputJsonPath": {
                    "type": "string"
                },
                "outputJsonPath": {
                    "type": "string"
                },
                "scenarioIndex": {
                    "type": "integer"
                },
                "stepIndex": {
                    "type": "integer"
                }
            }
        },
        "models.HTML": {
            "type": "object",
            "properties": {
                "ignoreBadJsonFiles": {
                    "type": "boolean"
                },
                "inputFolderPath": {
                    "type": "string"
                },
                "inputJsonPath": {
                    "type": "string"
                },
                "mergeFiles": {
                    "type": "boolean"
                },
                "metadata": {
                    "$ref": "#/definitions/models.Metadata"
                },
                "outputHtmlFolder": {
                    "type": "string"
                },
                "showEmbeddedFiles": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Metadata": {
            "type": "object",
            "properties": {
                "appVersion": {
                    "type": "string"
                },
                "browser": {
                    "type": "string"
                },
                "executed": {
                    "type": "string"
                },
                "parallel": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "testEnvironment": {
                    "type": "string"
                }
            }
        }
    }
}